templates/mssql-script/requirements.txt
cffi==1.16.0
cryptography==41.0.5
packaging==23.1
pycparser==2.21
pyodbc==5.0.1
pyOpenSSL==23.3.0
pyparsing==3.0.9
PyYAML==6.0.1
